requirements.txt
# Cancer Risk Assessor - Dependencies
# ==========================================

# API Framework
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9

# LLM & Embeddings (Google Vertex AI)
google-cloud-aiplatform==1.68.0
vertexai==1.68.0

# Vector Database
chromadb==1.3.7

# PDF Processing
marker-pdf==1.10.0  # Deep learning-based PDF extraction (superior for tables/layouts)
#pymupdf==1.23.21    # Required by marker as backend

# Data Validation & Settings
pydantic==2.9.0
pydantic-settings==2.2.1

# HTTP Client (for downloading PDF)
httpx==0.28.1

# Utilities
python-dotenv==1.0.1
tiktoken==0.6.0  # Token counting for chunking
tenacity==8.2.3  # Retry logic for API calls

# Development & Testing
pytest==7.4.4
pytest-asyncio==0.23.4

# Type hints
typing-extensions==4.11.0


sentence-transformers==5.2.2
chonkie==1.5.5

fitz
tools==1.0.23